requirements.txt
streamlit
cohere
scikit-learn
python-dotenv
